requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-Mail==0.9.1
Werkzeug==2.3.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
reportlab==4.0.4
qrcode==7.4.2
Pillow==10.0.0
itsdangerous==2.1.2
gunicorn==21.2.0